+++ src/txn_input.rs
@@ -1,3 +1,4 @@
+use bincode::{Decode, Encode};
 use chrono::Utc;
 use serde::Serialize;
 use serde_with::serde_as;
@@ -6,10 +7,10 @@ use std::collections::HashMap;
 use crate::{utils::output_map_to_bytes, wallet::Wallet};
 
 #[serde_as]
-#[derive(Debug, Clone, PartialEq, Serialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Decode, Encode)]
 pub struct TransactionInput {
-	pub timestamp: usize,
-	pub amount: usize,
+	pub timestamp: i64,
+	pub amount: u32,
 	#[serde_as(as = "serde_with::hex::Hex")]
 	pub sender_address: Vec<u8>,
 	#[serde_as(as = "serde_with::hex::Hex")]
@@ -19,9 +20,9 @@ pub struct TransactionInput {
 impl TransactionInput {
 	pub fn new(
 		sender_wallet: &Wallet,
-		output_map: &HashMap<Vec<u8>, usize>,
+		output_map: &HashMap<Vec<u8>, u32>,
 	) -> Self {
-		let timestamp = Utc::now().timestamp_millis() as usize;
+		let timestamp = Utc::now().timestamp_millis();
 		let output_bytes = output_map_to_bytes(&output_map);
 		let signature = sender_wallet
 			.sign(&output_bytes)
