+++ src/utils.rs
@@ -5,9 +5,9 @@ use sha3::{Digest, Sha3_256};
 pub fn cryptohash(
 	data: &[String],
 	last_hash: &str,
-	timestamp: usize,
-	nonce: usize,
-	difficulty: usize,
+	timestamp: i64,
+	nonce: u32,
+	difficulty: u32,
 ) -> Vec<u8> {
 	let mut hasher = Sha3_256::new();
 	let data = data.join(":");
@@ -17,7 +17,7 @@ pub fn cryptohash(
 	hasher.finalize().to_vec()
 }
 
-pub fn output_map_to_bytes(output_map: &HashMap<Vec<u8>, usize>) -> Vec<u8> {
+pub fn output_map_to_bytes(output_map: &HashMap<Vec<u8>, u32>) -> Vec<u8> {
 	let config = bincode::config::standard();
 	bincode::encode_to_vec(output_map, config)
 		.expect("Output bytes failed to encode.")
